+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to the "L13 Diff" extension will be documented in this file.
 
 ### Added
 - Added favorites explorer and activitybar icon.
-- Added an icon (★) in the top right of the diff panel to add a favorite diff.
+- Added an icon (★) in the top right of the diff panel to save a favorite diff.
 - Added context menu to open, rename or remove a favorite diff from the list.
 - Added setting `l13Diff.openFavoriteAndCompare` for clicking on a favorite diff in the list.
 - Added command `L13 Diff: Remove all favorites` to delete all favorite diffs.
-- Added predefined variable `${workspaceFolder}` for paths. 
+- Added predefined variables `${workspaceFolder}`, `${workspaceFolder:INDEX}` and `${workspaceFolderBasename:NAME}` for paths. 
+- Added keyboard shortcut `Cmd/Ctrl + D` to save current paths in favorites.
 
 ## [0.15.0] - 2019-06-09